# user-service/docs/swagger.yaml
openapi: 3.0.3
info:
  title: User Service API
  version: 1.0.0
  description: |
    REST API za korisnike, lokacije, favorite kompanije i rezervacije.
    Svi primjeri i status kodovi su uključeni. Autentikaciju trenutno ignorišemo.

servers:
  - url: http://localhost:3000/api
    description: Localhost

  - url: http://user-service:3000/api
    description: Docker network


tags:
  - name: health
    description: Health check
  - name: users
    description: Upravljanje korisnicima
  - name: locations
    description: Upravljanje lokacijama
  - name: favorites
    description: Favorite kompanije korisnika
  - name: reservations
    description: Upravljanje rezervacijama

paths:
  /health:
    get:
      tags: [health]
      summary: Health check
      responses:
        '200':
          description: Servis je živ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
              examples:
                ok:
                  value: { status: "ok" }

  /users:
    get:
      tags: [users]
      summary: Lista korisnika
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
              examples:
                sample:
                  value:
                    - { id: 1, name: "Ana", email: "ana@example.com" }
                    - { id: 2, name: "Marko", email: "marko@example.com" }
        '500': { $ref: '#/components/responses/ServerError' }

    post:
      tags: [users]
      summary: Kreiraj korisnika
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
            examples:
              sample:
                value: { name: "Ana", email: "ana@example.com" }
      responses:
        '201':
          description: Kreirano
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
              examples:
                sample:
                  value: { id: 10, name: "Ana", email: "ana@example.com" }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }

  /users/{id}:
    get:
      tags: [users]
      summary: Dohvati korisnika po ID-u
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
              examples:
                sample:
                  value: { id: 1, name: "Ana", email: "ana@example.com" }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

    put:
      tags: [users]
      summary: Ažuriraj korisnika
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdate' }
            examples:
              sample:
                value: { name: "Ana Ažurirana", email: "ana.new@example.com" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
              examples:
                sample:
                  value: { id: 1, name: "Ana Ažurirana", email: "ana.new@example.com" }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

    delete:
      tags: [users]
      summary: Obriši korisnika
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Obrisano (bez tijela)
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

  /locations:
    get:
      tags: [locations]
      summary: Lista lokacija
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Location' }
              examples:
                sample:
                  value:
                    - { id: 1, name: "Center", address: "Glavna 1" }
                    - { id: 2, name: "Novi Grad", address: "Sporedna 5" }
        '500': { $ref: '#/components/responses/ServerError' }

    post:
      tags: [locations]
      summary: Kreiraj lokaciju
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocationCreate' }
            examples:
              sample:
                value: { name: "Center", address: "Glavna 1" }
      responses:
        '201':
          description: Kreirano
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Location' }
              examples:
                sample:
                  value: { id: 3, name: "Center", address: "Glavna 1" }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }

  /locations/{id}:
    get:
      tags: [locations]
      summary: Dohvati lokaciju
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Location' }
              examples:
                sample:
                  value: { id: 1, name: "Center", address: "Glavna 1" }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

    put:
      tags: [locations]
      summary: Ažuriraj lokaciju
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocationUpdate' }
            examples:
              sample:
                value: { name: "Center 2", address: "Glavna 1A" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Location' }
              examples:
                sample:
                  value: { id: 1, name: "Center 2", address: "Glavna 1A" }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

    delete:
      tags: [locations]
      summary: Obriši lokaciju
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '204': { description: Obrisano }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

  /favorites/{userId}:
    get:
      tags: [favorites]
      summary: Lista favorita za korisnika
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Favorite' }
              examples:
                sample:
                  value:
                    - { id: 1, user_id: 5, company_name: "ACME d.o.o." }
                    - { id: 2, user_id: 5, company_name: "Globex" }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

  /favorites:
    post:
      tags: [favorites]
      summary: Dodaj u favorite
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FavoriteCreate' }
            examples:
              sample:
                value: { user_id: 5, company_name: "ACME d.o.o." }
      responses:
        '201':
          description: Kreirano
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Favorite' }
              examples:
                sample:
                  value: { id: 10, user_id: 5, company_name: "ACME d.o.o." }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }

    delete:
      tags: [favorites]
      summary: Ukloni iz favorita
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: integer }
        - name: favoriteId
          in: query
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Obrisano }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

  /reservations:
    get:
      tags: [reservations]
      summary: Lista rezervacija
      parameters:
        - name: limit
          in: query
          schema: { type: integer, example: 50 }
        - name: offset
          in: query
          schema: { type: integer, example: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Reservation' }
              examples:
                sample:
                  value:
                    - { id: 1, user_id: 5, service_id: 2, date: "2025-05-15" }
        '500': { $ref: '#/components/responses/ServerError' }

    post:
      tags: [reservations]
      summary: Kreiraj rezervaciju
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReservationCreate' }
            examples:
              sample:
                value: { user_id: 5, service_id: 2, date: "2025-05-15" }
      responses:
        '201':
          description: Kreirano
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reservation' }
              examples:
                sample:
                  value: { id: 21, user_id: 5, service_id: 2, date: "2025-05-15" }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }

  /reservations/{id}:
    get:
      tags: [reservations]
      summary: Dohvati rezervaciju
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reservation' }
              examples:
                sample:
                  value: { id: 1, user_id: 5, service_id: 2, date: "2025-05-15" }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

    put:
      tags: [reservations]
      summary: Ažuriraj rezervaciju
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReservationUpdate' }
            examples:
              sample:
                value: { user_id: 5, service_id: 3, date: "2025-05-20" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reservation' }
              examples:
                sample:
                  value: { id: 1, user_id: 5, service_id: 3, date: "2025-05-20" }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

    delete:
      tags: [reservations]
      summary: Obriši rezervaciju
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '204': { description: Obrisano }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
      description: Primarni ključ zapisa

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            sample:
              value: { error: "Validation failed" }

    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            sample:
              value: { error: "Resource not found" }

    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            sample:
              value: { error: "Unexpected server error" }

  schemas:
    Health:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Users
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
      required: [id, name, email]

    UserCreate:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
      required: [name, email]

    UserUpdate:
      allOf:
        - $ref: '#/components/schemas/UserCreate'

    # Locations
    Location:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        address: { type: string }
      required: [id, name]

    LocationCreate:
      type: object
      properties:
        name: { type: string }
        address: { type: string }
      required: [name]

    LocationUpdate:
      allOf:
        - $ref: '#/components/schemas/LocationCreate'

    # Favorites
    Favorite:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        company_name: { type: string }
      required: [id, user_id, company_name]

    FavoriteCreate:
      type: object
      properties:
        user_id: { type: integer }
        company_name: { type: string }
      required: [user_id, company_name]

    # Reservations
    Reservation:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        service_id: { type: integer }
        date: { type: string, format: date }
      required: [id, user_id, service_id, date]

    ReservationCreate:
      type: object
      properties:
        user_id: { type: integer }
        service_id: { type: integer }
        date: { type: string, format: date }
      required: [user_id, service_id, date]

    ReservationUpdate:
      allOf:
        - $ref: '#/components/schemas/ReservationCreate'